8th

+++ src/main/java/com/dynamodb/dto/CommentDto.java
@@ -5,7 +5,7 @@ import lombok.Data;
 
 @Data
 public class CommentDto {
-    private Integer commentId;
+    private Long commentId;
     private String commentator;
     private String description;
     private CommentType commentType;
